ref(onboarding): Move remaining js utils to js folder
