+++ static/app/gettingStartedDocs/javascript/astro/index.tsx
@@ -1,12 +1,12 @@
 import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
 import {featureFlag} from 'sentry/gettingStartedDocs/javascript/javascript/featureFlag';
+import {logsFullStack} from 'sentry/gettingStartedDocs/javascript/javascript/logs';
+import {profilingFullStack} from 'sentry/gettingStartedDocs/javascript/javascript/profiling';
 
 import {agentMonitoring} from './agentMonitoring';
 import {crashReport} from './crashReport';
 import {feedback} from './feedback';
-import {logs} from './logs';
 import {onboarding} from './onboarding';
-import {profiling} from './profiling';
 import {replay} from './replay';
 
 const docs: Docs = {
@@ -15,8 +15,17 @@ const docs: Docs = {
   replayOnboarding: replay,
   crashReportOnboarding: crashReport,
   featureFlagOnboarding: featureFlag,
-  logsOnboarding: logs,
-  profilingOnboarding: profiling,
+  logsOnboarding: logsFullStack({
+    docsPlatform: 'astro',
+    packageName: '@sentry/astro',
+  }),
+  profilingOnboarding: profilingFullStack({
+    packageName: '@sentry/astro',
+    browserProfilingLink:
+      'https://docs.sentry.io/platforms/javascript/guides/astro/profiling/browser-profiling/',
+    nodeProfilingLink:
+      'https://docs.sentry.io/platforms/javascript/guides/astro/profiling/node-profiling/',
+  }),
   agentMonitoringOnboarding: agentMonitoring,
 };
