--- static/app/gettingStartedDocs/javascript/ember/logs.tsx
@@ -1,9 +1,0 @@
-import {getJavascriptLogsOnboarding} from 'sentry/utils/gettingStartedDocs/javascript';
-
-import {installSnippetBlock} from './utils';
-
-export const logs = getJavascriptLogsOnboarding({
-  installSnippetBlock,
-  docsPlatform: 'ember',
-  packageName: '@sentry/ember',
-});
