--- static/app/gettingStartedDocs/javascript/gatsby/logs.tsx
@@ -1,9 +1,0 @@
-import {getJavascriptLogsOnboarding} from 'sentry/utils/gettingStartedDocs/javascript';
-
-import {installSnippetBlock} from './utils';
-
-export const logs = getJavascriptLogsOnboarding({
-  installSnippetBlock,
-  docsPlatform: 'gatsby',
-  packageName: '@sentry/gatsby',
-});
