+++ static/app/gettingStartedDocs/javascript/nextjs/index.tsx
@@ -1,13 +1,14 @@
 import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
 import {featureFlag} from 'sentry/gettingStartedDocs/javascript/javascript/featureFlag';
+import {logsFullStack} from 'sentry/gettingStartedDocs/javascript/javascript/logs';
+import {profilingFullStack} from 'sentry/gettingStartedDocs/javascript/javascript/profiling';
+import {tct} from 'sentry/locale';
 
 import {agentMonitoring} from './agentMonitoring';
 import {crashReport} from './crashReport';
 import {feedback} from './feedback';
-import {logs} from './logs';
 import {onboarding} from './onboarding';
 import {performance} from './performance';
-import {profiling} from './profiling';
 import {replay} from './replay';
 
 const docs: Docs = {
@@ -17,8 +18,69 @@ const docs: Docs = {
   performanceOnboarding: performance,
   crashReportOnboarding: crashReport,
   featureFlagOnboarding: featureFlag,
-  profilingOnboarding: profiling,
-  logsOnboarding: logs,
+  profilingOnboarding: profilingFullStack({
+    packageName: '@sentry/nextjs',
+    browserProfilingLink:
+      'https://docs.sentry.io/platforms/javascript/guides/nextjs/profiling/browser-profiling/',
+    nodeProfilingLink:
+      'https://docs.sentry.io/platforms/javascript/guides/nextjs/profiling/node-profiling/',
+    getProfilingHeaderContent: () => [
+      {
+        type: 'text',
+        text: tct(
+          'In Next.js you can configure document response headers via the headers option in [code:next.config.js]:',
+          {
+            code: <code />,
+          }
+        ),
+      },
+      {
+        type: 'code',
+        tabs: [
+          {
+            label: 'ESM',
+            language: 'javascript',
+            filename: 'next.config.js',
+            code: `
+  export default withSentryConfig({
+    async headers() {
+      return [{
+        source: "/:path*",
+        headers: [{
+          key: "Document-Policy",
+          value: "js-profiling",
+        }],
+      }];
+    },
+    // ... other Next.js config options
+  });`,
+          },
+          {
+            label: 'CJS',
+            language: 'javascript',
+            filename: 'next.config.js',
+            code: `
+  module.exports = withSentryConfig({
+    async headers() {
+      return [{
+        source: "/:path*",
+        headers: [{
+          key: "Document-Policy",
+          value: "js-profiling",
+        }],
+      }];
+    },
+    // ... other Next.js config options
+  });`,
+          },
+        ],
+      },
+    ],
+  }),
+  logsOnboarding: logsFullStack({
+    docsPlatform: 'nextjs',
+    packageName: '@sentry/nextjs',
+  }),
   agentMonitoringOnboarding: agentMonitoring,
 };
