--- static/app/gettingStartedDocs/javascript/nuxt/logs.tsx
@@ -1,6 +1,0 @@
-import {getJavascriptLogsFullStackOnboarding} from 'sentry/utils/gettingStartedDocs/javascript';
-
-export const logs = getJavascriptLogsFullStackOnboarding({
-  docsPlatform: 'nuxt',
-  packageName: '@sentry/nuxt',
-});
