--- static/app/gettingStartedDocs/javascript/react-router/logs.tsx
@@ -1,6 +1,0 @@
-import {getJavascriptLogsFullStackOnboarding} from 'sentry/utils/gettingStartedDocs/javascript';
-
-export const logs = getJavascriptLogsFullStackOnboarding({
-  docsPlatform: 'react-router',
-  packageName: '@sentry/react-router',
-});
