+++ static/app/gettingStartedDocs/javascript/react/index.tsx
@@ -1,13 +1,14 @@
 import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
 import {featureFlag} from 'sentry/gettingStartedDocs/javascript/javascript/featureFlag';
-import {logs} from 'sentry/gettingStartedDocs/javascript/react/logs';
+import {logs} from 'sentry/gettingStartedDocs/javascript/javascript/logs';
+import {profiling} from 'sentry/gettingStartedDocs/javascript/javascript/profiling';
 
 import {crashReport} from './crashReport';
 import {feedback} from './feedback';
 import {onboarding} from './onboarding';
 import {performance} from './performance';
-import {profiling} from './profiling';
 import {replay} from './replay';
+import {installSnippetBlock} from './utils';
 
 const docs: Docs = {
   onboarding,
@@ -15,8 +16,16 @@ const docs: Docs = {
   replayOnboarding: replay,
   performanceOnboarding: performance,
   crashReportOnboarding: crashReport,
-  profilingOnboarding: profiling,
-  logsOnboarding: logs,
+  profilingOnboarding: profiling({
+    installSnippetBlock,
+    docsLink:
+      'https://docs.sentry.io/platforms/javascript/guides/react/profiling/browser-profiling/',
+  }),
+  logsOnboarding: logs({
+    installSnippetBlock,
+    docsPlatform: 'react',
+    packageName: '@sentry/react',
+  }),
   featureFlagOnboarding: featureFlag,
 };
