--- static/app/gettingStartedDocs/javascript/solidstart/logs.tsx
@@ -1,6 +1,0 @@
-import {getJavascriptLogsFullStackOnboarding} from 'sentry/utils/gettingStartedDocs/javascript';
-
-export const logs = getJavascriptLogsFullStackOnboarding({
-  docsPlatform: 'solidstart',
-  packageName: '@sentry/solidstart',
-});
