+++ static/app/gettingStartedDocs/javascript/tanstackstart-react/index.tsx
@@ -1,15 +1,24 @@
 import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {logsFullStack} from 'sentry/gettingStartedDocs/javascript/javascript/logs';
+import {profilingFullStack} from 'sentry/gettingStartedDocs/javascript/javascript/profiling';
 
 import {agentMonitoring} from './agentMonitoring';
-import {logs} from './logs';
 import {onboarding} from './onboarding';
-import {profiling} from './profiling';
 
 const docs: Docs = {
   onboarding,
-  profilingOnboarding: profiling,
+  profilingOnboarding: profilingFullStack({
+    packageName: '@sentry/tanstackstart-react',
+    browserProfilingLink:
+      'https://docs.sentry.io/platforms/javascript/guides/tanstackstart-react/profiling/browser-profiling/',
+    nodeProfilingLink:
+      'https://docs.sentry.io/platforms/javascript/guides/tanstackstart-react/profiling/node-profiling/',
+  }),
   agentMonitoringOnboarding: agentMonitoring,
-  logsOnboarding: logs,
+  logsOnboarding: logsFullStack({
+    docsPlatform: 'tanstackstart-react',
+    packageName: '@sentry/tanstackstart-react',
+  }),
 };
 
 export default docs;
