--- static/app/gettingStartedDocs/javascript/tanstackstart-react/logs.tsx
@@ -1,6 +1,0 @@
-import {getJavascriptLogsFullStackOnboarding} from 'sentry/utils/gettingStartedDocs/javascript';
-
-export const logs = getJavascriptLogsFullStackOnboarding({
-  docsPlatform: 'tanstackstart-react',
-  packageName: '@sentry/tanstackstart-react',
-});
